# backend/config/api-spec.yaml
openapi: 3.0.0
info:
  title: Gloire Road Map API
  version: 1.0.0
  description: A comprehensive goal tracking system with attendance monitoring
  contact:
    name: Gloire Road Map Team
    email: support@gloire-roadmap.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3005
    description: Development server
  - url: https://api.gloire-roadmap.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token obtained from /api/auth/login
  schemas:
    Student:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          description: Unique student identifier
          example: 1
        name:
          type: string
          description: Full name of the student
          example: John Doe
        contact_number:
          type: string
          nullable: true
          description: Student contact phone number
          example: "+1-234-567-8900"
        address:
          type: string
          nullable: true
          description: Student physical address
          example: "123 Main St, City, State 12345"
        date_of_birth:
          type: string
          format: date
          nullable: true
          description: Student date of birth
          example: "2000-01-15"
        points:
          type: integer
          description: Achievement points earned
          example: 45
        created_at:
          type: string
          format: date-time
          description: Registration timestamp
          example: "2025-10-01T00:00:00.000Z"
        days_attended:
          type: integer
          description: Number of days marked as present
          example: 18
        total_attendance_records:
          type: integer
          description: Total attendance records
          example: 20
    Goal:
      type: object
      required:
        - student_id
        - title
      properties:
        id:
          type: integer
          description: Unique goal identifier
          example: 1
        student_id:
          type: integer
          description: Reference to student
          example: 1
        title:
          type: string
          description: Goal title
          example: Complete Math Assignment
        description:
          type: string
          nullable: true
          description: Detailed goal description
          example: Complete chapters 1-5 of algebra textbook
        target_date:
          type: string
          format: date
          nullable: true
          description: Target completion date
          example: "2025-10-15"
        is_completed:
          type: boolean
          description: Completion status
          example: false
        completed_at:
          type: string
          format: date-time
          nullable: true
          description: Actual completion timestamp
          example: "2025-10-14T14:30:00.000Z"
        created_at:
          type: string
          format: date-time
          description: Goal creation timestamp
          example: "2025-10-01T09:00:00.000Z"
    Attendance:
      type: object
      required:
        - student_id
        - date
        - status
      properties:
        id:
          type: integer
          description: Unique attendance record identifier
          example: 1
        student_id:
          type: integer
          description: Reference to student
          example: 1
        date:
          type: string
          format: date
          description: Attendance date
          example: "2025-10-07"
        status:
          type: string
          enum: [present, absent, late, excused]
          description: Attendance status
          example: present
        notes:
          type: string
          nullable: true
          description: Optional notes about attendance
          example: Arrived 5 minutes late due to traffic
        created_at:
          type: string
          format: date-time
          description: Record creation timestamp
          example: "2025-10-07T08:00:00.000Z"
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: Resource not found
        error:
          type: string
          description: Error type
          example: NotFoundError
        details:
          type: string
          description: Additional error details
          example: Student with ID 999 does not exist
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          description: Success message
          example: Operation completed successfully
security:
  - bearerAuth: []
...
paths:
...
  /api/students:
    get:
      tags:
        - Students
      summary: Get all students
      description: Retrieves all students with their attendance statistics
      responses:
        "200":
          description: List of students retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
        "401":
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - Students
      summary: Create new student
      description: Creates a new student record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: John Doe
                contact_number:
                  type: string
                  example: "+1-234-567-8900"
                address:
                  type: string
                  example: "123 Main St, City, State 12345"
                date_of_birth:
                  type: string
                  format: date
                  example: "2000-01-15"
      responses:
        "201":
          description: Student created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/students/{id}:
    get:
      tags:
        - Students
      summary: Get student by ID
      description: Retrieves a specific student by their ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Student ID
      responses:
        "200":
          description: Student retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      tags:
        - Students
      summary: Update student
      description: Updates an existing student record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Student ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                contact_number:
                  type: string
                address:
                  type: string
                date_of_birth:
                  type: string
                  format: date
      responses:
        "200":
          description: Student updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags:
        - Students
      summary: Delete student
      description: Deletes a student record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Student ID
      responses:
        "200":
          description: Student deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/students/{id}/goals:
    get:
      tags:
        - Students
      summary: Get student goals
      description: Retrieves all goals for a specific student
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Student ID
      responses:
        "200":
          description: Student goals retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - Students
      summary: Create goal for student
      description: Creates a new goal for a specific student
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Student ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  example: Complete Math Assignment
                description:
                  type: string
                  example: Complete chapters 1-5 of algebra textbook
                target_date:
                  type: string
                  format: date
                  example: "2025-10-15"
      responses:
        "201":
          description: Goal created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/attendance/student/{student_id}:
    get:
      tags:
        - Attendance
      summary: Get attendance for specific student
      description: Retrieves attendance records for a specific student
      parameters:
        - in: path
          name: student_id
          required: true
          schema:
            type: integer
          description: Student ID
        - in: query
          name: start_date
          schema:
            type: string
            format: date
          description: Filter from start date (YYYY-MM-DD)
        - in: query
          name: end_date
          schema:
            type: string
            format: date
          description: Filter to end date (YYYY-MM-DD)
      responses:
        "200":
          description: Student attendance records retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...